chore(bgatlas): extract into separate module
