+++ client/scripts/atlas/index.js
@@ -1,0 +1,4 @@
+require('./module')
+var modulename = require('./modulename')
+
+module.exports = modulename
