+++ client/scripts/configs/states.js
@@ -557,29 +557,6 @@ module.config(/* @ngInject */function ($locationProvider, $stateProvider, $urlRo
         }
       })
   })
-
-  /// ////////
-  // Atlas //
-  /// ////////
-  $stateProvider
-    .state('auth.atlas', {
-      url: '/atlas',
-      views: {
-        'content@auth': {
-          templateUrl: 'views/atlas/dashboard.html',
-          controller: 'AtlasDashboardController as $ctrl'
-        }
-      }
-    })
-    .state('auth.atlas.request', {
-      url: '/request',
-      views: {
-        'content@auth': {
-          templateUrl: 'views/atlas/request.html',
-          controller: 'AtlasRequestController as $ctrl'
-        }
-      }
-    })
 })
   .run(/* @ngInject */function ($rootScope, $state, $stateParams, authorization, user) {
     $rootScope.$state = $state
